added validations on UI and backend

+++ routes/application.js
@@ -7,9 +7,28 @@ let Permission = models.Permission;
 let Application = models.Application;
 let AppPermissions = models.App_Permissions;
 let Sequelize = require('sequelize');
+const { body, query, check, validationResult, oneOf } = require('express-validator');
 const Op = Sequelize.Op;
+const errorFormatter = ({ location, msg, param, value, nestedErrors }) => {    
+  return `${msg}`;
+};
 
-router.post('/', function(req, res, next) {
+
+router.post('/', [
+  body('name')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Application Name'),
+  body('email').optional({checkFalsy:true})
+    .isEmail().withMessage('Invalid Email Address'),
+  body('owner')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Owner Name'),
+  body('clientId')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Client Id'),    
+], (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  }
+
 	let appObj = req.body;
   Application.create({
     "name": appObj.name,
@@ -26,7 +45,21 @@ router.post('/', function(req, res, next) {
   
 });
 
-router.put('/', function(req, res, next) {
+router.put('/', [
+  body('name')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Application Name'),
+  body('email').optional({checkFalsy:true})
+    .isEmail().withMessage('Invalid Email Address'),
+  body('owner')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Owner Name'),
+  body('clientId')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Client Id'),    
+], (req, res) => {
+  const errors = validationResult(req).formatWith(errorFormatter);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  }
+
 	let appObj = req.body;
 	console.log(appObj);
   Application.update({ 
@@ -53,7 +86,14 @@ router.get('/all', function(req, res, next) {
   
 });
 
-router.get('/', function(req, res, next) {
+router.get('/', oneOf([
+    query('id').isUUID(4).withMessage("Invalid Id"),
+    query('id').isInt().withMessage("Invalid Id"),
+]), (req, res) => {
+  const errors = validationResult(req).formatWith(errorFormatter);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  } 
 	let appId = req.query.id;
   Application.findOne({where:{"id":appId}}).then((result) => {  	  	
 	  res.json(result);
@@ -62,7 +102,14 @@ router.get('/', function(req, res, next) {
   });
 });
 
-router.delete('/', function(req, res, next) {
+router.delete('/', oneOf([
+    query('id').isUUID(4).withMessage("Invalid Id"),
+    query('id').isInt().withMessage("Invalid Id"),
+]), (req, res) => {
+  const errors = validationResult(req).formatWith(errorFormatter);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  } 
   let appId = req.query.id;
   Application.destroy({where:{"id":appId}}).then((application) => {        
     res.json({"result":"success"});
