+++ routes/roles.js
@@ -3,6 +3,10 @@ var router = express.Router();
 let models = require('../models');
 let PermissionTemplate = models.PermissionTemplate;
 let Role = models.Role;
+const { body, query, check, validationResult, oneOf } = require('express-validator');
+const errorFormatter = ({ location, msg, param, value, nestedErrors }) => {    
+  return `${location}[${param}]: ${msg}`;
+};
 
 router.get('/permissions', function(req, res, next) {
   PermissionTemplate.findAll({where:{"applicationType": req.query.applicationType}}).then((result) => {  	  	
@@ -13,7 +17,16 @@ router.get('/permissions', function(req, res, next) {
   
 });
 
-router.post('/', function(req, res, next) {
+router.post('/', [
+  body('name')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Role Name'),
+  body('managedBy')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Managed By')    
+], (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  }
 	let roleObj = req.body;
   Role.create({
     "name": roleObj.name,
@@ -25,11 +38,19 @@ router.post('/', function(req, res, next) {
   	res.json({"result":"success"});
   }).catch(function(err) {
      console.log("error occured: "+err);
-  });
-  
+  });  
 });
 
-router.put('/', function(req, res, next) {
+router.put('/', [
+  body('name')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Role Name'),
+  body('managedBy')
+    .matches(/^[a-zA-Z]{1}[a-zA-Z0-9_-]*$/).withMessage('Invalid Managed By')    
+], (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  }
 	let roleObj = req.body;	
   Role.update({ 
   	"name": roleObj.name,
@@ -52,7 +73,14 @@ router.get('/all', function(req, res, next) {
   });  
 });
 
-router.get('/', function(req, res, next) {
+router.get('/', oneOf([
+    query('id').isUUID(4).withMessage("Invalid Id"),
+    query('id').isInt().withMessage("Invalid Id"),
+]), (req, res) => {
+  const errors = validationResult(req).formatWith(errorFormatter);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  } 
 	let roleId = req.query.id;
   Role.findOne({where:{"id":roleId}}).then((result) => {  	  	
 	  res.json(result);
@@ -61,7 +89,14 @@ router.get('/', function(req, res, next) {
   });
 });
 
-router.delete('/', function(req, res, next) {
+router.delete('/', oneOf([
+    query('id').isUUID(4).withMessage("Invalid Id"),
+    query('id').isInt().withMessage("Invalid Id"),
+]), (req, res) => {
+  const errors = validationResult(req).formatWith(errorFormatter);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ success: false, errors: errors.array() });
+  } 
   let roleId = req.query.id;
   Role.destroy({where:{"id":roleId}}).then((role) => {        
     res.json({"result":"success"});
